# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Devices Service
tags:
  - name: Devices
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /swagger:
    x-swagger-pipe: swagger_raw

  /rest/services/v1/apps/devices/manufacturers:
    get:
      tags:
        - Devices
      summary: Returns a list of manufacturers
      operationId: getManufacturerList
      consumes:
        - application/json
      responses:
        200:
          description: List of manufacturers retrieved successfully.
          schema:
            $ref: '#/definitions/ManufacturerList-Model'
        400:
          description: Missing required input parameter.
        404:
          description: No manufactures found.
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure.
        500:
          description: Technical error occured, service failure.

  /rest/services/v1/apps/devices/models:
    get:
      tags:
        - "Devices"
      operationId: getDeviceModels
      summary: Returns all device models for a manufacturer
      consumes:
        - application/json
      parameters:
        - in: "query"
          name: manufacturer
          type: array
          items:
            type: string
          required: false
      responses:
        200:
          description: All device models retrieved successfully
          schema:
            $ref: "#/definitions/DeviceModel-Model"
        400:
          description: Missing required input parameter
        404:
          description: No device models found
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure

# complex objects have schema definitions
definitions:
  # Get Manufacturer List
  ManufacturerList-Model:
    type: object
    properties:
      result:
        type: object
        properties:
          brandName:
            type: array
            items:
              type: string

  # Get Device Models
  DeviceModel-Model:
    type: object
    properties:
      messages:
        type: array
        items:
          type: string
      result:
        type: array
        items:
          $ref: "#/definitions/ModelDetails"
      successful:
        type: boolean
      code:
        type: integer

  # Device model details
  ModelDetails:
    type: object
    properties:
      manufacturer:
        type: string
      model:
        type: string